pyproject.toml
[project]
name = "autoproger"
version = "2.0.0"
description = "Multi-agent system for automated software development and maintenance"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.14",
    "aiosqlite>=0.20",
    "asyncpg>=0.30",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "anthropic>=0.42",
    "httpx>=0.28",
    "gitpython>=3.1",
    "PyGithub>=2.5",
    "python-gitlab>=5.1",
    "python-telegram-bot>=21",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.18",
    "rich>=13",
    "tenacity>=9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6",
    "ruff>=0.8",
]

[tool.setuptools.packages.find]
include = ["core*", "agents*", "llm*", "integrations*", "context*", "quality*", "storage*", "api*", "bot*"]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
